pyproject.toml
[tool.poetry]
name = "heroes-replay"
version = "0.1.0"
description = ""
authors = ["Ice Breaker Engineer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
SQLAlchemy = "^2.0.35"
Flask-SQLAlchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.7"
python-slugify = "^8.0.4"
flask-admin = "^1.6.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
